Code cleanup, added more environment variables & direct spark deployment.
Deleted various unused import statements, updated logging code, added
environment variables for optional namespace deletion/cleanup.
Added option to deploy spark pods without the need for oshinko.
Jgrafzahl tests now work under any namespace.

+++ test-blockchainanalysis/src/test/java/com/redhat/xpaas/Setup.java
@@ -1,27 +1,16 @@
 package com.redhat.xpaas;
 
 import com.redhat.xpaas.logger.LogWrapper;
-import com.redhat.xpaas.logger.LoggerUtil;
 import com.redhat.xpaas.openshift.OpenshiftUtil;
-import com.redhat.xpaas.oshinko.api.OshinkoWebUI;
-import com.redhat.xpaas.oshinko.deployment.Oshinko;
 import com.redhat.xpaas.rad.BlockChainAnalysis.api.BlockChainAnalysisSparkWebUI;
 import com.redhat.xpaas.rad.BlockChainAnalysis.api.BlockChainAnalysisWebUI;
 import com.redhat.xpaas.util.Tuple;
-import com.redhat.xpaas.wait.WaitUtil;
-import io.fabric8.kubernetes.api.model.Pod;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.concurrent.TimeoutException;
-import java.util.function.BooleanSupplier;
-import java.util.function.Predicate;
 
 import static com.redhat.xpaas.rad.BlockChainAnalysis.deployment.BlockChainAnalysis.deployBlockChainAnalysis;
 import static com.redhat.xpaas.rad.BlockChainAnalysis.deployment.BlockChainAnalysisSpark.deployBlockChainAnalysisSpark;
 
 public class Setup {
-  LogWrapper log = new LogWrapper(Setup.class, "blockchain");
+  private LogWrapper log = new LogWrapper(Setup.class, "blockchain");
   private String NAMESPACE = RadConfiguration.masterNamespace();
   private static final OpenshiftUtil openshift = OpenshiftUtil.getInstance();
   private static BlockChainAnalysisWebUI BlockChainAnalysis;
@@ -43,11 +32,13 @@ public class Setup {
       log.action("shutting-down-blockchainanalysisspark-webdrivers", () -> BlockChainAnalysisSpark.webDriverCleanup());
     }
 
-    log.action("deleting-namespace", () -> openshift.deleteProject(NAMESPACE));
+    if(RadConfiguration.deleteNamespaceAfterTests()){
+      log.action("deleting-namespace", () -> openshift.deleteProject(NAMESPACE));
+    }
   }
 
   private void initializeProject(){
-    OpenshiftUtil.getInstance().createProject(NAMESPACE, true);
+    OpenshiftUtil.getInstance().createProject(NAMESPACE, RadConfiguration.recreateNamespace());
   }
 
 }
